Setup new-home to be the new home route
