+++ app/home/-components/hero-banner/styles.scss
