+++ app/home/-components/integrations-section/styles.scss
