+++ app/home/-components/support-section/template.hbs
@@ -7,7 +7,7 @@
     <div local-class='flex-container'>
         <div local-class='flex-row'>
             <div>
-                <NewHome::-Components::SupportSection::SupportItem
+                <Home::-Components::SupportSection::SupportItem
                     data-test-support-search
                     local-class='supportItem'
                     @icon='search'
@@ -22,7 +22,7 @@
                 >
             </div>
             <div>
-                <NewHome::-Components::SupportSection::SupportItem
+                <Home::-Components::SupportSection::SupportItem
                     data-test-support-design
                     local-class='supportItem'
                     @icon='pencil-alt'
@@ -39,7 +39,7 @@
         </div>
         <div local-class='flex-row'>
             <div>
-                <NewHome::-Components::SupportSection::SupportItem
+                <Home::-Components::SupportSection::SupportItem
                     data-test-support-analyze
                     local-class='supportItem'
                     @icon='chart-area'
@@ -53,7 +53,7 @@
                     alt={{t 'new-home.support-section.arrow'}}
                 >
             </div>
-            <NewHome::-Components::SupportSection::SupportItem
+            <Home::-Components::SupportSection::SupportItem
                 data-test-support-publish
                 @icon='file-alt'
                 @header={{t 'new-home.support-section.publish.header'}}
@@ -61,7 +61,7 @@
             />
         </div>
     </div>
-    <NewHome::-Components::SupportSection::LearnMoreButton
+    <Home::-Components::SupportSection::LearnMoreButton
         data-test-learn-more-button
         local-class='learnMoreButton'
     />
