--- app/new-home/controller.ts
@@ -1,10 +1,0 @@
-import Controller from '@ember/controller';
-
-export default class NewHome extends Controller {
-}
-
-declare module '@ember/controller' {
-    interface Registry {
-        'new-home': NewHome;
-    }
-}
