--- app/new-home/template.hbs
@@ -1,4 +1,0 @@
-<NewHome::-Components::HeroBanner />
-<NewHome::-Components::SupportSection />
-<NewHome::-Components::Testimonials />
-<NewHome::-Components::IntegrationsSection />
