+++ app/router.ts
@@ -95,11 +95,10 @@ const Router = EmberRouter.extend({
 Router.map(function() {
     // All non-guid routes (except error routes) belong above "Guid Routing"
     this.route('home', { path: '/' });
-    this.route('new-home');
+    this.route('old-home');
     this.route('dashboard');
     this.route('goodbye');
     this.route('institutions');
-    this.route('new-home');
     this.route('quickfiles');
     this.route('register');
     this.route('settings', function() {
